ai_service/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
numpy==1.26.0
pandas==2.1.0
scikit-learn==1.3.0
joblib==1.3.0
mysql-connector-python==8.2.0
requests==2.31.0